config.py
# app/config.py

import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    DEBUG = os.getenv("FLASK_DEBUG", "0") == "1"
    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URL")
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SECRET_KEY = os.getenv("SECRET_KEY", "dev_secret")
    WHISPER_URL = os.getenv("WHISPER_URL")
    WHISPER_API_KEY = os.getenv("WHISPER_API_KEY")
